+++ src/data/directory_content.rs
@@ -324,7 +324,15 @@ impl DirectoryContent {
         &self.state
     }
 
-    /// Returns one directory entry by index
+    /// Returns an iterator in the given range of the directory contents.
+    /// No filters are applied using this iterator.
+    pub fn iter_range_mut(
+        &mut self,
+        range: std::ops::Range<usize>,
+    ) -> impl Iterator<Item = &mut DirectoryEntry> {
+        self.content[range].iter_mut()
+
+        /// Returns one directory entry by index
     pub fn get(&mut self, i: usize) -> Option<&mut DirectoryEntry> {
         self.content.get_mut(i)
     }